Cleaning code
